+++ braintree/transaction_search.py
@@ -70,6 +70,11 @@ class TransactionSearch:
         Constants.get_all_constant_values_from_class(CreditCard.CustomerLocation)
     )
 
+    debit_network = Search.MultipleValueNodeBuilder(
+        "debit_network",
+        Constants.get_all_enum_values(CreditCard.DebitNetwork)
+    )
+
     source = Search.MultipleValueNodeBuilder("source")
 
     status = Search.MultipleValueNodeBuilder(