+++ braintree/util/constants.py
@@ -3,3 +3,5 @@ class Constants(object):
     @staticmethod
     def get_all_constant_values_from_class(klass):
         return [klass.__dict__[item] for item in dir(klass) if not item.startswith("__")]
+    def get_all_enum_values(enum_class):
+        return [item.value for item in enum_class]