+++ tests/test_helper.py
@@ -86,19 +86,20 @@ def showwarning(*_):
 warnings.showwarning = showwarning
 
 class TestHelper(object):
+    adyen_merchant_account_id = "adyen_ma"
+    aib_swe_ma_merchant_account_id = "aib_swe_ma"
+    another_us_bank_merchant_account_id = "another_us_bank_merchant_account"
+    card_processor_brl_merchant_account_id = "card_processor_brl"
     default_merchant_account_id = "sandbox_credit_card"
+    fake_amex_direct_merchant_account_id = "fake_amex_direct_usd"
+    fake_first_data_merchant_account_id = "fake_first_data_merchant_account"
+    fake_venmo_account_merchant_account_id = "fake_first_data_venmo_account"
+    hiper_brl_merchant_account_id = "hiper_brl"
     non_default_merchant_account_id = "sandbox_credit_card_non_default"
     non_default_sub_merchant_account_id = "sandbox_sub_merchant_account"
+    pinless_debit_merchant_account_id = "pinless_debit"
     three_d_secure_merchant_account_id = "three_d_secure_merchant_account"
-    fake_amex_direct_merchant_account_id = "fake_amex_direct_usd"
-    fake_venmo_account_merchant_account_id = "fake_first_data_venmo_account"
-    fake_first_data_merchant_account_id = "fake_first_data_merchant_account"
     us_bank_merchant_account_id = "us_bank_merchant_account"
-    another_us_bank_merchant_account_id = "another_us_bank_merchant_account"
-    adyen_merchant_account_id = "adyen_ma"
-    hiper_brl_merchant_account_id = "hiper_brl"
-    card_processor_brl_merchant_account_id = "card_processor_brl"
-    aib_swe_ma_merchant_account_id = "aib_swe_ma"
 
     add_on_discount_plan = {
          "description": "Plan for integration tests -- with add-ons and discounts",